shared/offer: move useBookOfferModal as it will be used in favorite page
